driver: rename and move DigitalOutputPowerDriver
The DigitalOutputPowerDriver was primarly meant for a device where the power
cycle took very long due to an internal battery. Since we have the ResetProtocol
now, move and refactor the driver to only implement the ResetProtocol. This way
PowerProtocol and ResetProtocol can be used side by side.
